pyproject.toml
[project]
name = "gcp-icons-for-plantuml"
version = "0.1.0"
description = "A collection of Google Cloud Platform icons for PlantUML"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "pillow>=11.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[project.scripts]
gcp-icons-for-plantuml = "gcp_icons_for_plantuml.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
